Initial import of sunset MyNews (CMS) project.
This project is legacy/sunset and only available for archival/historical purposes.

I worked on this from ~1999-2006 or so before ultimately migrating over to and
contributing to the Drupal project.  This code ran alienated.org in the early years
as well as running a bunch of internal Intranet sites at the 800lbGorilla that I worked
for until ~2006.

Unfortunately, I lost the original SVN trunk for this, so all versioning and iterations are
lost to the ifnar.

ChangeLog is in place, so perhaps info can be gleaned from there.